% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cantons}
\alias{cantons}
\title{cantons}
\format{
\subsection{\code{cantons}}{

A sf object with 26 rows and 2 columns:
\describe{
\item{NAME_1}{Name of Canton}
\item{geometry}{polygon coordinates}
}
}
}
\source{
GADM database \url{https://gadm.org/}
}
\usage{
cantons
}
\description{
Geometries of Cantons of Switzerland. CRS is unassigned, but should be EPSG:3857.
}
\keyword{datasets}
