% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH_migration_data}
\alias{CH_migration_data}
\title{CH_migration_data}
\format{
\subsection{\code{CH_migration_data}}{

A data frame with 325 rows and 8 columns:
\describe{
\item{id_a, id_b}{Names of Cantons A and B}
\item{flow_ab}{Number of migrations from A to B}
\item{flow_ba}{Number of migrations from B to A}
\item{xa,ya}{Longitude and latitude of the centroid of Canton A. Web-Mercator projection (EPSG: 3857)}
\item{xb,yb}{Longitude and latitude of the centroid of Canton B. Web-Mercator projection (EPSG: 3857)}
}
}
}
\source{
Federal Statistical Office of Switzerland, under OPEN-BY-ASK terms of use: \url{https://www.bfs.admin.ch/bfs/de/home/statistiken/bevoelkerung/migration-integration/binnenwanderung.assetdetail.3222163.html}
}
\usage{
CH_migration_data
}
\description{
Internal migrations between Cantons of Switzerland, 2011-2016.
}
\keyword{datasets}
