% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlegend.R
\name{flowlegend}
\alias{flowlegend}
\title{flowlegend}
\usage{
flowlegend(
  x,
  pos = NULL,
  title = "Title 1",
  title2 = NULL,
  col = "#CCCCCC",
  txtcol = "black"
)
}
\arguments{
\item{x}{flow generetd with the function plotflows(). list.}

\item{pos}{position of the legend.}

\item{title}{title of the legend (flows), default is "Title".}

\item{title2}{title of the legend (circles), default is NULL.}

\item{col}{color of the legend, default is #CCCCC}

\item{txtcol}{color of the text, default is black}
}
\description{
Build flowmaps legend
}
\examples{
library(sf)
crs <- "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
subregions <- st_read(system.file("subregions.gpkg", package="flowmapper")) \%>\% st_transform(crs)
migr <- read.csv(system.file("migrantstocks2019.csv", package="flowmapper"))
threshold <- 1500
migr <- migr[migr$fij >= threshold,]
intra <- migr[migr$i == migr$j,c("i","fij")]
colnames(intra) <- c("id","nb")

plot(st_geometry(subregions), col = "#CCCCCC", border = "white", lwd = 0.5)
flows <- flowmap(
 x = subregions,
 xid = "id",
 df = migr,
 dfid = c("i","j"),
 dfvar = "fij"
)

flowlegend(x = flows, title = "inter", title2 = "intra")
}
