% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap_sf.R
\name{flowmap_sf}
\alias{flowmap_sf}
\title{Create sf for flowmap}
\usage{
flowmap_sf(
  flowdat = NULL,
  od = NULL,
  nodes = NULL,
  crs = NULL,
  outline_linewidth = 0.01,
  alpha = 0.8,
  nodes_alpha = 0.8,
  outline_col = "black",
  k_nodes = NULL,
  node_buffer_factor = 1.2,
  node_radius_factor = 1,
  edge_offset_factor = 1,
  node_fill_factor = NULL,
  edge_width_factor = 1.2,
  arrow_point_angle = 45
)
}
\arguments{
\item{flowdat}{Input dataframe. See details below.}

\item{od}{As an alternative to \code{flowdat}, dataframe with the origin-destination pairs and the flow between them.  Must contain the columns o, d, value. \code{nodes} must be provided as well. See details below.}

\item{nodes}{As an alternative to \code{flowdat}, a dataframe with the nodes of the network. Must contain the columns name, x, y. See details below.}

\item{crs}{A spatial reference system (CRS) code of tha coordinates passed in \code{flowdat} or \code{od}.}

\item{outline_linewidth}{The linewidth of the outline of the arrows.}

\item{alpha}{Opacity of the edges.}

\item{nodes_alpha}{Opacity of the nodes.}

\item{outline_col}{Color of the outline of the edges.}

\item{k_nodes}{Number of clusters to group nodes into. If defined, nodes will be clustered hierarchically based on spatial proximity. By default, no clustering will be applied.}

\item{node_buffer_factor}{Controls the distance between the nodes and the edges ( in multiple of the nodes' radii).}

\item{node_radius_factor}{Controls the size of the nodes.}

\item{edge_offset_factor}{Controls the distance between the parallel arrows.}

\item{node_fill_factor}{Controls the downscaling of the fill of the nodes ( as to not outshine the edges ).}

\item{edge_width_factor}{Controls the width of the edges.}

\item{arrow_point_angle}{Controls the pointiness of the edges.}
}
\value{
The ggplot with an additional polygon layer for the flow arrows and an additional polygon layer for the nodes
}
\description{
Create sf for flowmap
}
\details{
The function requires as inputs a dataframe \code{flowdat} which contains for every combination of two nodes a and b the coordinates of these nodes as well as the intensity of flow between those nodes in both directions (a to b, b to a). The dataframe should have the following columns:
\itemize{
\item \strong{id_a:} The unique id of node a
\item \strong{id_b:} The unique id of node b
\item \strong{xa:} The x coordinate of node a
\item \strong{ya:} The y coordinate of node a
\item \strong{xb:} The x coordinate of node b
\item \strong{yb:} The y coordinate of node b
\item \strong{flow_ab:} The intensity of flow from node a to node b
\item \strong{flow_ba:} The intensity of flow from node b to node a
}

Alternatively, the function can take as input a dataframe \code{od} which contains the origin-destination pairs and the flow between them. The dataframe should have the following columns:
\itemize{
\item \strong{o:} The unique id of the origin node
\item \strong{d:} The unique id of the destination node
\item \strong{value:} The intensity of flow between the origin and destination
}
In this case, the function also requires a dataframe \code{nodes} which contains the coordinates of the nodes. The dataframe should have the following columns:
\itemize{
\item \strong{name:} The unique id of the node
\item \strong{x:} The x coordinate of the node
\item \strong{y:} The y coordinate of the node
}

The function will impose coord_equal() on the ggplot.

Inspired by \href{https://flowmap.gl/}{flowmap.gl}.
}
\examples{
data <- flowmapper::CH_migration_data
cantons <- flowmapper::cantons
sf::st_crs(cantons) <- 3857
flows_nodes <- flowmap_sf(flowdat = data, nodes = cantons, crs = 3857)
}
