% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmapper-package.R
\docType{package}
\name{flowmapper-package}
\alias{flowmapper}
\alias{flowmapper-package}
\title{flowmapper: Draw Flows (Migration, Goods, Money, Information) on 'ggplot2' Plots}
\description{
Adds flow maps to 'ggplot2' plots. The flow maps consist of 'ggplot2' layers which visualize the nodes as circles and the bilateral flows between the nodes as bidirectional half-arrows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JohMast/flowmapper}
  \item Report bugs at \url{https://github.com/JohMast/flowmapper/issues}
}

}
\author{
\strong{Maintainer}: Johannes Mast \email{johannes.mast@dlr.de} (\href{https://orcid.org/0000-0001-6595-5834}{ORCID})

}
\keyword{internal}
