% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flowmap.R
\name{get_circle_coords}
\alias{get_circle_coords}
\title{Helper function to create coordinates for circles of nodes}
\usage{
get_circle_coords(center = c(0, 0), r = 1, npoints = 25)
}
\arguments{
\item{center}{center y and y coordinates}

\item{r}{radius}

\item{npoints}{number of points}
}
\value{
a dataframe with x and y coordinates of the circle
}
\description{
Helper function to create coordinates for circles of nodes
}
\author{
Johannes Mast, Credit to https://stackoverflow.com/a/6863490
}
