% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_flowdat.R
\name{hca_flowdat}
\alias{hca_flowdat}
\title{Use hierarchical clustering to merge nodes based on proximity}
\usage{
hca_flowdat(flowdat, k = 20)
}
\arguments{
\item{flowdat}{The data containing flows from a to b, b to a, and the coordinates of a and b}

\item{k}{The number of nodes to keep.}
}
\value{
a dataframe of the same format as flowdat, but with some nodes (and their flows) merged. Note that this will in most cases contain some circular flows (a to a) even if the input flowdat did not.
}
\description{
Use hierarchical clustering to merge nodes based on proximity
}
