% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flowmap.R
\name{short_scale}
\alias{short_scale}
\title{Create short scale format for numbers in the legend}
\usage{
short_scale(x, digits = 3)
}
\arguments{
\item{x}{The number}

\item{digits}{Significant digits}
}
\description{
Create short scale format for numbers in the legend
}
\author{
Johannes Mast, credit: https://stackoverflow.com/a/59086755
}
