% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flowmap.R
\name{util_data_flow_to_flowdat}
\alias{util_data_flow_to_flowdat}
\title{util_data_flow_to_flowdat}
\usage{
util_data_flow_to_flowdat(nodes, flows)
}
\arguments{
\item{nodes}{A data frame with the nodes of the network}

\item{flows}{A data frame with the flow data}
}
\value{
A data frame with the flow data in flowdat format
}
\description{
This function takes a flow data frame in long format and a data frame with the nodes coordinates and returns a flowdat data frame
}
\details{
Helper function to merge od data in long data and nodes to flowdat format
}
\examples{
#nodes <- data.frame(name=c("a","b","c"),x=c(0,1,2),y=c(0,1,2))
#flow <- data.frame(o=c("a","b"),d=c("b","c"),value=c(1,2))
#util_data_flow_to_flowdat(nodes,flow)
}
\author{
Johannes Mast,
}
